Seminar/anti_kuba_bot/requirements.txt
discord.py
translators